# Improve SectionedResults performance 