--- a/src/realm/mixed.cpp
+++ b/src/realm/mixed.cpp
@@ -617,7 +617,12 @@ size_t Mixed::hash() const
             hash = murmur2_or_cityhash(unsigned_data, 12);
             break;
         }
-        case type_Decimal:
+        case type_Decimal: {
+            auto value = get<Decimal128>();
+            auto unsigned_data = reinterpret_cast<const unsigned char*>(value.raw());
+            hash = murmur2_or_cityhash(unsigned_data, sizeof(Decimal128::Bid128));
+            break;
+        }
         case type_Mixed:
         case type_Link:
         case type_LinkList: