--- a/src/realm/object-store/sectioned_results.hpp
+++ b/src/realm/object-store/sectioned_results.hpp
@@ -22,6 +22,7 @@
 #include <realm/object-store/results.hpp>
 
 #include <list>
+#include <unordered_map>
 
 namespace realm {
 class Mixed;
@@ -31,13 +32,12 @@ struct SectionedResultsChangeSet;
 
 /// For internal use only. Used to track the indices for a given section.
 struct Section {
-    Section() = default;
     size_t index = 0;
     Mixed key;
     std::vector<size_t> indices;
 };
 
-using SectionedResultsNotificatonCallback = util::UniqueFunction<void(SectionedResultsChangeSet)>;
+using SectionedResultsNotificationCallback = util::UniqueFunction<void(SectionedResultsChangeSet)>;
 
 /**
  * An instance of ResultsSection gives access to elements in the underlying collection that belong to a given section
@@ -50,7 +50,7 @@ using SectionedResultsNotificatonCallback = util::UniqueFunction<void(SectionedR
  */
 class ResultsSection {
 public:
-    ResultsSection();
+    ResultsSection() = default;
 
     /// Retrieve an element from the section for a given index.
     Mixed operator[](size_t idx) const;
@@ -74,13 +74,13 @@ public:
      *
      * @param callback The function to execute when a insertions, modification or deletion in this `ResultsSection`
      * was detected.
-     * @param key_path_array A filter that can be applied to make sure the `SectionedResultsNotificatonCallback` is
+     * @param key_path_array A filter that can be applied to make sure the `SectionedResultsNotificationCallback` is
      * only executed when the property in the filter is changed but not otherwise.
      *
      * @return A `NotificationToken` that is used to identify this callback. This token can be used to remove the
      * callback via `remove_callback`.
      */
-    NotificationToken add_notification_callback(SectionedResultsNotificatonCallback callback,
+    NotificationToken add_notification_callback(SectionedResultsNotificationCallback&& callback,
                                                 std::optional<KeyPathArray> key_path_array = std::nullopt) &;
 
     bool is_valid() const;
@@ -89,7 +89,7 @@ private:
     friend class SectionedResults;
     ResultsSection(SectionedResults* parent, Mixed key);
 
-    SectionedResults* m_parent;
+    SectionedResults* m_parent = nullptr;
     Mixed m_key;
     std::unique_ptr<char[]> m_key_buffer;
     Section* get_if_valid() const;
@@ -100,18 +100,18 @@ private:
  * An instance of `SectionedResults` allows access to elements from underlying `Results` collection
  * where elements are arranged into sections defined by a key either from a user defined sectioning algorithm
  * or a predefined built-in sectioning algorithm. Elements are then accessed through a `ResultsSection` which can be
- * retreived through the subscript operator on `SectionedResults`.
+ * retrieved through the subscript operator on `SectionedResults`.
  */
 class SectionedResults {
 public:
     SectionedResults() = default;
-    using SectionKeyFunc = util::UniqueFunction<Mixed(Mixed value, std::shared_ptr<Realm> realm)>;
+    using SectionKeyFunc = util::UniqueFunction<Mixed(Mixed value, const std::shared_ptr<Realm>& realm)>;
 
     /**
      * Returns a `ResultsSection` which will be bound to a section key present at the given index in
      * `SectionedResults`.
      *
-     * NOTE: A `ResultsSection` is lazily retreived, meaning that the index it was retreived from
+     * NOTE: A `ResultsSection` is lazily retrieved, meaning that the index it was retreived from
      * is not guaranteed to be the index of this `ResultsSection` at the time of access.
      * For example if this `ResultsSection` is at index 1 and the  `ResultsSection`
      * below this one is deleted, this `ResultsSection` will now be at index 0
@@ -133,13 +133,13 @@ public:
      *
      * @param callback The function to execute when a insertions, modification or deletion in this `SectionedResults`
      * was detected.
-     * @param key_path_array A filter that can be applied to make sure the `SectionedResultsNotificatonCallback` is
+     * @param key_path_array A filter that can be applied to make sure the `SectionedResultsNotificationCallback` is
      * only executed when the property in the filter is changed but not otherwise.
      *
      * @return A `NotificationToken` that is used to identify this callback. This token can be used to remove the
      * callback via `remove_callback`.
      */
-    NotificationToken add_notification_callback(SectionedResultsNotificatonCallback callback,
+    NotificationToken add_notification_callback(SectionedResultsNotificationCallback&& callback,
                                                 std::optional<KeyPathArray> key_path_array = std::nullopt) &;
 
     /// Return a new instance of SectionedResults that uses a snapshot of the underlying `Results`.
@@ -159,43 +159,35 @@ public:
 
 private:
     friend class Results;
+    friend class realm::ResultsSection;
+
     /// SectionedResults should not be created directly and should only be instantiated from `Results`.
     SectionedResults(Results results, SectionKeyFunc section_key_func);
-    SectionedResults(Results results, Results::SectionedResultsOperator op, util::Optional<StringData> prop_name);
-
-    /// Used for creating a frozen or snapshot of SectionedResults.
-    SectionedResults(Results&& results, std::map<Mixed, Section>&& sections,
-                     std::map<size_t, Mixed>&& current_section_index_to_key_lookup,
-                     std::list<std::string>&& current_str_buffers)
-        : m_has_performed_initial_evalutation(true)
-        , m_results(std::move(results))
-        , m_sections(std::move(sections))
-        , m_current_section_index_to_key_lookup(std::move(current_section_index_to_key_lookup))
-        , m_current_str_buffers(std::move(current_str_buffers))
-    {
-    }
+    SectionedResults(Results results, Results::SectionedResultsOperator op, StringData prop_name);
 
     friend struct SectionedResultsNotificationHandler;
     util::CheckedOptionalMutex m_mutex;
     SectionedResults copy(Results&&) REQUIRES(!m_mutex);
     void calculate_sections_if_required() REQUIRES(m_mutex);
     void calculate_sections() REQUIRES(m_mutex);
-    bool m_has_performed_initial_evalutation = false;
+    bool m_has_performed_initial_evaluation = false;
     NotificationToken
-    add_notification_callback_for_section(Mixed section_key, SectionedResultsNotificatonCallback callback,
+    add_notification_callback_for_section(Mixed section_key, SectionedResultsNotificationCallback&& callback,
                                           std::optional<KeyPathArray> key_path_array = std::nullopt);
 
-    friend class realm::ResultsSection;
     Results m_results;
     SectionKeyFunc m_callback;
-    std::map<Mixed, Section> m_sections GUARDED_BY(m_mutex);
+    std::vector<Section> m_sections GUARDED_BY(m_mutex);
+
     // Returns the key of the current section from its index.
-    std::map<size_t, Mixed> m_current_section_index_to_key_lookup GUARDED_BY(m_mutex);
+    // Returns the key of the previous section from its index.
+    std::vector<Mixed> m_previous_index_to_key;
+
     // Stores the Key, Section Index of the previous section
     // so we can efficiently calculate the collection change set.
-    std::map<Mixed, size_t> m_previous_key_to_index_lookup;
-    // Returns the key of the previous section from its index.
-    std::map<size_t, Mixed> m_prev_section_index_to_key;
+    std::unordered_map<Mixed, size_t> m_current_key_to_index;
+    std::unordered_map<Mixed, size_t> m_previous_key_to_index;
+
     // By passing the index of the object from the underlying `Results`,
     // this will give a pair with the section index of the object, and the position of the object in that section.
     // This is used for parsing the indices in CollectionChangeSet to section indices.
@@ -211,11 +203,11 @@ private:
 
 struct SectionedResultsChangeSet {
     /// Sections and indices in the _new_ collection which are new insertions
-    std::map<size_t, IndexSet> insertions;
+    std::vector<IndexSet> insertions;
     /// Sections and indices of objects in the _old_ collection which were modified
-    std::map<size_t, IndexSet> modifications;
+    std::vector<IndexSet> modifications;
     /// Sections and indices which were removed from the _old_ collection
-    std::map<size_t, IndexSet> deletions;
+    std::vector<IndexSet> deletions;
     /// Indexes of sections which are newly inserted.
     IndexSet sections_to_insert;
     /// Indexes of sections which are deleted from the _old_ collection.