--- a/test/object-store/set.cpp
+++ b/test/object-store/set.cpp
@@ -705,8 +705,6 @@ TEMPLATE_TEST_CASE("set", "[set]", CreateNewSet<void>, ReuseSet<void>)
             CHECK(set().insert(-1).second);
         });
 
-        auto x = set().min();
-
         REQUIRE(set().is_valid());
         CHECK(set().sum(col_int_set) == 578);
         CHECK(set().min(col_int_set) == -1);