… src/realm/object-store/sectioned_results.cpp
… test/object-store/sectioned_results.cpp