A first pass at addressing https://github.com/realm/realm-swift/issues/8087.

SectionedResults used a std::map in a few places where the keys are a dense range (i.e. always [0..map.size())) and so they can be std::vector instead. The maps keyed on Mixed are now std::unordered_map.

Change notifications now report changes as a `std::vector<IndexSet>` rather than `std::map<size_t, IndexSet>`. This is slower and uses more memory when the only sections that changed are near the end of a SectionedResults with a large number of sections, but is much faster if all sections changed or if the sections which changed are early in the SectionedResults. Change notifications now reuse buffers, which increases persistent memory usage slightly but significant reduces allocations.

Change notifications for a single section now only compute the changes for that section rather than computing the full changes and then filtering out the changes for other sections.

There's a pile of minor optimizations here and there. Eliminating redundant TableRef checks speeds things up by a few percent. The new main thread scheduler is *extremely* minor: it makes thread checking nearly twice as fast, but thread checking was already a tiny part of the runtime. It makes literally everything faster on the thread where performance matters the most, though.

After these changes >80% of the runtime is spent in the sectioning callback function. This means there's probably not really much room for further improvement when using user-defined sectioning functions, but maybe we can do something clever when sectioning on a key path (as the reporting user is doing).

Benchmark changes:

| Test | Section Count | Before | After | % Speedup |
| ---- | ------------- | ------ | ----- | --------- |
| create and get section count | 1 | 10.2938 | 9.61949 | 7 |
| create and get section count | 10 | 12.4977 | 9.58887 | 30 |
| create and get section count | 1000 | 15.5668 | 9.86488 | 58 |
| create and get section count | 10000 | 23.6001 | 11.3862 | 107 |
| iterate directly | 1 | 7.17427 | 5.54021 | 29 |
| iterate directly | 10 | 9.92673 | 7.57087 | 31 |
| iterate directly | 1000 | 66.8946 | 62.5418 | 7 |
| iterate directly | 10000 | 72.0806 | 63.9037 | 13 |
| iterate over a snapshot | 1 | 11.2224 | 10.4199 | 8 |
| iterate over a snapshot | 10 | 12.0267 | 10.5125 | 14 |
| iterate over a snapshot | 1000 | 16.5787 | 12.4366 | 33 |
| iterate over a snapshot | 10000 | 23.5323 | 13.3865 | 76 |
| change notification | 1 | 27.9859 | 25.1588 | 11 |
| change notification | 10 | 30.8994 | 25.234 | 22 |
| change notification | 1000 | 40.3322 | 25.5394 | 58 |
| change notification | 10000 | 69.6117 | 29.165 | 139 |
| single section change notification | 1 | 28.2577 | 25.1852 | 12 |
| single section change notification | 10 | 31.0225 | 23.8774 | 30 |
| single section change notification | 1000 | 40.2219 | 24.1318 | 67 |
| single section change notification | 10000 | 68.6834 | 25.7545 | 167 |

"iterate over a snapshot" becomes faster than "iterate directly" because benchmark happens to be sectioning a Table directly and the sectioning chosen results in an extremely pessimal access pattern for directly indexing a Table, so constructing a TableView first actually becomes faster.