@tgoyne 2023-05-12
tsan failure looks real but unrelated